backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "black>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
]

[tool.black]
line-length = 79
target-version = ["py311"]

[flake8]
max-line-length = 80
extend-ignore = ["E203", "W503"]
